refactor(tests): share the OpenAPI path-item reader and tidy the new providers
Review follow-ups on #861:

- `operationMethods()` had grown a third identical copy, so the list of HTTP
  method keys — the one thing deciding what counts as an operation rather than a
  sibling property — was maintained in three places. Extracted to
  `Support\OpenApiPathItemTrait` and used by all three `Schemas/OpenApi*Test`
  classes.
- `bareePathsWithACanonicalTwin` → `barePathsWithACanonicalTwin` (typo), and the
  two deprecation providers iterate `array_keys()` rather than binding a path
  item they never read.
