+++ phpunit_tests/Schemas/OpenApiDataRiteSegmentTest.php
@@ -8,6 +8,7 @@ use LiturgicalCalendar\Api\Enum\PathCategory;
 use LiturgicalCalendar\Api\Enum\Rite;
 use LiturgicalCalendar\Api\Http\Exception\ValidationException;
 use LiturgicalCalendar\Api\Params\RegionalDataParams;
+use LiturgicalCalendar\Tests\Support\OpenApiPathItemTrait;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
@@ -27,6 +28,8 @@ use PHPUnit\Framework\TestCase;
  */
 final class OpenApiDataRiteSegmentTest extends TestCase
 {
+    use OpenApiPathItemTrait;
+
     /** @var array<string,mixed> */
     private static array $openapi;
 
@@ -245,25 +248,5 @@ final class OpenApiDataRiteSegmentTest extends TestCase
         }
 
         self::assertGreaterThan(0, $seen, 'openapi.json documents no /data calendar-definition path at all');
-    }
-
-    /**
-     * The HTTP methods a path item documents, sorted, with any non-method key (`parameters`,
-     * `summary`, `description`, `servers`) filtered out — those are legal siblings of the operations
-     * in an OpenAPI path item and are not operations themselves.
-     *
-     * @param array<string,mixed> $pathItem
-     *
-     * @return list<string>
-     */
-    private static function operationMethods(array $pathItem): array
-    {
-        $methods = array_values(array_intersect(
-            array_keys($pathItem),
-            ['get', 'put', 'post', 'delete', 'options', 'head', 'patch', 'trace']
-        ));
-        sort($methods);
-
-        return $methods;
     }
 }
